dedupe-tile-component * Merge all duplicated tile components and styling * Add variant prop to Tile to select which variant to use
