--- src/dashboards/Compact/components/Tile/Tile.scss
@@ -1,44 +1,0 @@
-.compact .tile {
-    height: 100%;
-    width: 25rem;
-    background-color: var(--tavla-box-background-color);
-    flex: none;
-    padding: 2rem;
-    overflow: scroll;
-
-    &--pressed {
-        box-shadow: inset 0 0 65px -20px rgba(0, 0, 0, 0.54);
-    }
-
-    &__header {
-        display: flex;
-        justify-content: space-between;
-        align-items: center;
-        margin-bottom: 1rem;
-
-        h2 {
-            margin: 0;
-            color: var(--tavla-font-color);
-            overflow: hidden;
-            white-space: nowrap;
-            text-overflow: ellipsis;
-        }
-
-        svg {
-            margin-left: 0.5rem;
-        }
-    }
-
-    &__header-icons {
-        display: flex;
-        align-items: center;
-        font-size: 2rem;
-    }
-
-    &__walking-time {
-        font-style: normal;
-        font-weight: normal;
-        font-size: 1rem;
-        color: var(--tavla-font-color);
-    }
-}
